requirements.txt
appdirs==1.4.3
Django==1.11.1
django-extensions==1.8.1
django-extra-views==0.9.0
django-recurrence==1.4.1
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
python-http-client==3.0.0
pytz==2017.2
sendgrid==3.6.5
sendgrid-django==4.0.4
six==1.10.0
